Changes file structure to separate out user caching logic and export from index
