+++ src/server/api/errors.js
@@ -1,6 +1,6 @@
 import { GraphQLError } from 'graphql/error'
 import { r } from '../models'
-import { userHasRole } from '../models/cacheable-queries'
+import { userHasRole } from '../models/cacheable_queries'
 
 const accessHierarchy = ['TEXTER', 'SUPERVOLUNTEER', 'ADMIN', 'OWNER']
